Infinity_CRM/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
crispy-tailwind==0.5.0
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==6.1.0
django-rest-framework==0.1.0
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.2.0
Pillow
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
static3==0.7.0
tk==0.1.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
